Reorganize the mesh tests and mesh test library
